requirements/dev.txt
-r test.txt
fastapi[all]
uvicorn[standard]
black
motor
python-jose
bcrypt
passlib
